refactor(cli): extract lint module to reduce file size under 1500 lines
Move lint-related functions (run_lint_cmd, collect_skill_dirs,
lint_report_to_json, etc.) from cli.rs into cli/lint.rs to satisfy
the file size gate. cli.rs was 1534 lines; now 1424 lines.

- New file: crates/dcc-mcp-cli/src/presentation/cli/lint.rs
- Remove unused imports (BTreeSet, IssueSeverity, dcc_mcp_skills types)
- Update Command::Lint arm to call lint::run_lint_cmd

Fixes file-size gate failure in PR #1992.
